% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{append_by_enrich}
\alias{append_by_enrich}
\title{Add vertices to a list based on p-value cutoff / enrichment}
\usage{
append_by_enrich(G, v1, v2, cutoff, collect = FALSE)
}
\arguments{
\item{G}{igraph which contains vertex identifiers present in `v1` and `v2`.}

\item{v1}{vector of vertex identifiers. This list will be added to from `v2`
if there are any vertices passing the enrichment criteria.}

\item{v2}{vector of vertex identifiers to potentially add to `v1`.}

\item{cutoff}{numeric. Cutoff enrichment value (p-value) for which to add
vertices from `v2` to `v1` the list.}

\item{collect}{logical. If TRUE, the output will contain a listed object with
the names of `adjusted_list` and `remaining_vertices`.}
}
\description{
Add vertices to a list based on p-value cutoff / enrichment
}
\author{
Christopher Nobles, Ph.D.
}
