% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{append_vertices_to_lists}
\alias{append_vertices_to_lists}
\title{Append vertices in a list to vertex lists depending on enrichment}
\usage{
append_vertices_to_lists(G, L, v, cutoff, collect = FALSE)
}
\arguments{
\item{G}{igraph which contains vertex identifiers present in `L` and `v`.}

\item{L}{list of vectors with vertex identifiers. This list will function as
the base lists for which vertices from `v` will be added to based on the best
enrichment.}

\item{v}{vector of vertex identifiers to be added to vertex lists in `L`.
Vertices will only be added if they are below the `cutoff` p-value. If there
are multiple possibilities, vertices will only be added to the most
significant match.}

\item{cutoff}{numeric. Cutoff enrichment value (p-value) for which to add
vertices from `v` to vectors in `L`.}

\item{collect}{logical. If TRUE, the output will contain a listed object with
the names of `adjusted_lists` and `remaining_vertices`.}
}
\description{
Append vertices in a list to vertex lists depending on enrichment
}
\author{
Christopher Nobles, Ph.D.
}
