% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_prob_vertex}
\alias{calc_prob_vertex}
\title{Approximate the probability of each vertex within a list belonging to the
list.}
\usage{
calc_prob_vertex(v, G)
}
\arguments{
\item{v}{vector of vertex identifiers. Vertices must be present in graph G.}

\item{G}{igraph object representing the graph and including vertices from v.}
}
\description{
Approximate the probability of each vertex within a list belonging to the
list.
}
\author{
Christopher Nobles, Ph.D.
}
