% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_bipartite_prob}
\alias{comp_bipartite_prob}
\title{Approximate the probability of interactions between two subgraphs, a
bipartite comparison.}
\usage{
comp_bipartite_prob(v1, v2, G, exact = TRUE)
}
\arguments{
\item{v1}{vector of vertex identifiers, indicating the vertices present in
subgraph 1. Vertex identifiers need to be the names of vertices in G. "v1"
vertices will be used to create g1 subgraph.}

\item{v2}{vector of vertex identifiers. Same criteria as L1. L1 and L2 for
comparison should not have zero intersection. Vertices in both L1 and L2 will
be removed during analysis. "v2" vertices will be used to create g2 subgraph.}

\item{G}{igraph object of which g1 and g2 subgraphs will be created from.}

\item{exact}{logical. If TRUE, the approximation of the probability is
calculated rather than approximated by `ppois()`.}
}
\description{
Approximate the probability of interactions between two subgraphs, a
bipartite comparison.
}
\author{
Christopher Nobles, Ph.D.
}
