% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_edgeset_prob}
\alias{comp_edgeset_prob}
\title{Compute edgeset probability}
\usage{
comp_edgeset_prob(el, G, maxEdgeCnt, elcols = c(1, 2))
}
\arguments{
\item{el}{data.frame/matrix edgelist where column 1 and column 2 contain the 
lists of connected vertices of interest.}

\item{G}{igraph graph of which vertices from `el` are present.}

\item{maxEdgeCnt}{numeric maximum number of edges that could be obsersered 
given vertices within edgelist.}

\item{elcols}{numeric vector of length 2, indicating the two columns of the 
input data.frame or matrix (`el`) which contain the edge list data, defaults 
to 1 and 2.}
}
\description{
Compute edgeset probability
}
\author{
Christopher Nobles, Ph.D.
}
