% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_lambda}
\alias{comp_lambda}
\title{Compute the number of expected connections between a set of vertices.}
\usage{
comp_lambda(v, G)
}
\arguments{
\item{v}{vector of verticies present in graph G.}

\item{G}{igraph object.}
}
\description{
Compute the number of expected connections between a set of vertices.
}
\author{
Christopher Nobles, Ph.D.
}
