% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_lambda_bipartite}
\alias{comp_lambda_bipartite}
\title{Compute the number of expected connections between two sets of vertices.}
\usage{
comp_lambda_bipartite(v1, v2, G)
}
\arguments{
\item{v1}{vector of vertex identifiers for list 1.}

\item{v2}{vector of vertex identifiers for list 2.}

\item{G}{igraph of which vertices from v1 and v2 are present.}
}
\description{
Compute the number of expected connections between two sets of vertices.
}
\author{
Christopher Nobles, Ph.D.
}
