% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_outside_vertex_prob}
\alias{comp_outside_vertex_prob}
\title{Approximate the probability of a vertex not within a list belonging to the
list.}
\usage{
comp_outside_vertex_prob(v1, v2, G, exact = TRUE)
}
\arguments{
\item{v1}{vector of vertex identifiers. Vertices must be present in graph G.}

\item{v2}{vector of vertex identifiers not in `v1`. These vertices will be
tested as a vertex in `v1`.}

\item{G}{igraph object representing the graph and including vertices from
`v1` and `v2`.}

\item{exact}{logical. If TRUE, the approximation of the probability is
calculated rather than approximated by `ppois()`.}
}
\description{
Approximate the probability of a vertex not within a list belonging to the
list.
}
\author{
Christopher Nobles, Ph.D.
}
