% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_pij}
\alias{comp_pij}
\title{Compute the probability of vertex i connecting to vertex j}
\usage{
comp_pij(ki, kj, M, sparseOnly = TRUE)
}
\arguments{
\item{ki}{integer degree of vertex i}

\item{kj}{integer degree of vertex j}

\item{M}{integer size of graph G, the total number of edges within G.}
}
\description{
Compute the probability of vertex i connecting to vertex j
}
\author{
Christopher Nobles, Ph.D.
}
