% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_sparse_prob}
\alias{comp_sparse_prob}
\title{Comput probability for sparse graph edges.}
\usage{
comp_sparse_prob(lambda, edgeCnt, maxEdgeCnt, exact = TRUE)
}
\description{
Comput probability for sparse graph edges.
}
