% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{count_edges}
\alias{count_edges}
\title{Count edges within a graph or edges between two non-overlapping graphs}
\usage{
count_edges(G, v1, v2 = NULL, edges = FALSE)
}
\arguments{
\item{G}{igraph which contains vertex identified by `v1` and `v2`.}

\item{v1}{vector of vertex identifiers.}

\item{v2}{vector of vertex identifiers. `NULL` by default, but when a second
vector is supplied, the fuction will return the number of edges connecting
the two subgraphs indicated by the vertex lists.}

\item{edges}{logical. If TRUE, the edges rather than the count will be
returned as a data.frame with two columns named "vi" and "vj".}
}
\description{
Count edges within a graph or edges between two non-overlapping graphs
}
\author{
Christopher Nobles, Ph.D.
}
