% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{enrich_edgeset}
\alias{enrich_edgeset}
\title{Approximate the probability of observing an edge set}
\usage{
enrich_edgeset(v, G)
}
\arguments{
\item{v}{vector of vertex identifiers. Vertices must be present in graph G.}

\item{G}{igraph object representing the graph and including vertices from v.}
}
\description{
Approximate the probability of observing an edge set
}
\author{
Christopher Nobles, Ph.D.
}
