% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{enrich_outside_vertex}
\alias{enrich_outside_vertex}
\title{Approximate the probability of a vertex not within a list belonging to the
list}
\usage{
enrich_outside_vertex(v1, v2, G)
}
\arguments{
\item{v1}{vector of vertex identifiers. Vertices must be present in graph G.}

\item{v2}{vector of vertex identifiers not in `v1`. These vertices will be
tested as a vertex in `v1`.}

\item{G}{igraph object representing the graph and including vertices from
`v1` and `v2`.}
}
\description{
Approximate the probability of a vertex not within a list belonging to the
list
}
\author{
Christopher Nobles, Ph.D.
}
