% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{map_names}
\alias{map_names}
\title{Match gene names within data to vertex identifiers with a reference data set.}
\usage{
map_names(data, names = "gene_name", refdata, refcol = "preferred_name",
  refout = "protein_external_id", outname = "stringID",
  removeUnmapped = TRUE, quiet = FALSE)
}
\arguments{
\item{data}{data.frame with column specified by the `names` parameter.}

\item{names}{character specifying the column name in data containing the 
`names` of to be mapped from.}

\item{refdata}{data.frame with reference data. Needs to contain the columns
referred to by `refcol` and `refout`.}

\item{refcol}{character the name of the column with matching information for
the `names` parameter.}

\item{refout}{character the name of the column for in the reference with the 
information to be mapped to data.}

\item{outname}{character the name given to the new information mapped to the 
`data` data.frame.}

\item{removeUnmapped}{logical indicating if rows should be removed if
unmapped.}

\item{quiet}{logical indicating if messages should be printed or not.}
}
\description{
Match gene names within data to vertex identifiers with a reference data set.
}
\author{
Christopher Nobles, Ph.D.
}
