% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bipartite}
\alias{plot_bipartite}
\title{Plot bipartite graph using ggnet2}
\usage{
plot_bipartite(v1, v2, G, ..., quiet = FALSE)
}
\arguments{
\item{v1}{vector of vertex identifiers for subgraph g1. All vertex
identifiers need to be present in the graph G.}

\item{v2}{vector of vertex identifiers for subgraph g2. All vertex
identifiers need to be present in the graph G.}

\item{G}{igraph object of which v1 and v2 vertices will be used to create a
subgraph.}

\item{...}{arguments to be passed into ggnet2.}
}
\description{
Plot bipartite graph using ggnet2
}
\author{
Christopher Nobles, Ph.D.
}
