% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot graph with a list of vertex lists using ggnet2}
\usage{
plot_clusters(Lc, G, set.v = "Set1", set.e = "grey50", ...)
}
\arguments{
\item{Lc}{list of character vectors holding information on vertex ids for
different clusters or groups of vertices. Each object in the list will be
given a different color. Additionally, names of the list will be used in the
legend, unless no names are supplied. If a vertex in named on more than one
list, it will be colored black and pushed into an intersecting group.}

\item{G}{igraph object from which vertices in Lc will be used to create a
subgraph.}

\item{set.v}{character, Set name to be passed on to RColorBrewer::brewer.pal
to select color palette for vertices. If name supplied is not in
brewer.pal.info, then the name will be treated as a color and passed as a
constant color for vertices.}

\item{set.e}{character, Set name to be passed on to RColorBrewer::brewer.pal
to select color palette for edges. If name supplied is not in
brewer.pal.info, then the name will be treated as a color and passed as a
constant color for edges.}

\item{...}{arguments to be passed into ggnet2.}
}
\description{
Plot graph with a list of vertex lists using ggnet2
}
\author{
Christopher Nobles, Ph.D.
}
