% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_subgraph}
\alias{plot_subgraph}
\title{Plot subgraph using ggnet2}
\usage{
plot_subgraph(v, G, ...)
}
\arguments{
\item{v}{vector of vertex identifiers. All vertex identifiers need to be
present in the graph G.}

\item{G}{igraph object of which v vertices will be used to create a subgraph.}

\item{...}{arguments to be passed into ggnet2.}
}
\description{
Plot subgraph using ggnet2
}
\author{
Christopher Nobles, Ph.D.
}
