% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{prune_by_enrich}
\alias{prune_by_enrich}
\title{Prune vertices from vertex list based on p-value cutoff / enrichment}
\usage{
prune_by_enrich(G, v, cutoff, collect = FALSE)
}
\arguments{
\item{G}{igraph which contains vertex identifiers present in `v`.}

\item{v}{vector of vertex identifiers to be considered for pruning.}

\item{cutoff}{numeric. Cutoff enrichment value (p-value) for which to prune
vertices from the list.}

\item{collect}{logical. If TRUE, the output will contain a listed object with
the names of `pruned_list` and `removed_vertices`.}
}
\description{
Prune vertices from vertex list based on p-value cutoff / enrichment
}
\author{
Christopher Nobles, Ph.D.
}
