% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_edgeset_prob}
\alias{comp_edgeset_prob}
\title{Compute edgeset probability.}
\usage{
comp_edgeset_prob(v, G, exact = TRUE)
}
\arguments{
\item{v}{vector of vertex identifiers. Vertices must be present in graph G.}

\item{G}{igraph object representing the graph and including vertices from v.}

\item{exact}{logical. If TRUE, the approximation of the probability is
calculated rather than approximated by `ppois()`.}

\item{maxEdgeCnt}{numeric maximum number of edges that could be obsersered
given vertices within edgelist.}
}
\description{
Compute edgeset probability.
}
\author{
Christopher Nobles, Ph.D.
}
