% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{comp_lambda}
\alias{comp_lambda}
\title{Compute the number of expected connections between a set of vertices}
\usage{
comp_lambda(el, G, elcols = c(1, 2))
}
\arguments{
\item{el}{data.frame/matrix edgelist where column 1 and column 2 contain the 
lists of connected vertices of interest.}

\item{G}{igraph graph of which vertices from `el` are present.}

\item{elcols}{numeric vector of length 2, indicating the two columns of the 
input data.frame or matrix (`el`) which contain the edge list data, defaults 
to 1 and 2.}
}
\description{
Compute the number of expected connections between a set of vertices
}
\author{
Christopher Nobles, Ph.D.
}
