% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{construct_graph}
\alias{construct_graph}
\title{Construct graph from data.frame}
\usage{
construct_graph(df, V = NULL, dfcols = c(1, 2), mode = "undirected")
}
\arguments{
\item{df}{data.frame where the initial two columns represent an edgelist.}

\item{V}{vector of vertex names / identifiers. Vertices not present in the 
edgelist will be included as vertices with no edges in the output graph. If 
nothing supplied, only vertices present in edgelist data.frame will be
present in the graph.}

\item{dfcols}{numeric vector of length 2, specifying the columns of the 
data.frame that contain the edgelist information.}

\item{mode}{character. Indicates if the graph is a directed or undirected}
}
\description{
Construct graph from data.frame
}
\author{
Christopher Nobles, Ph.D.
}
